requirements.txt
scikit_image==0.16.2
ipdb==0.12.2
opencv_python==4.2.0.32
imageio==2.6.1
torchvision==0.4.0a0+9232c4a
h5py==2.7.1
numpy==1.18.0
torch==1.2.0
matplotlib==2.1.2
visdom==0.1.8.9
moviepy==1.0.3
Pillow==7.1.2
scikit_learn==0.23.0
skimage==0.0
tensorboardX==2.0
